refactor: remove ipfs on path
