--- src/ipfs-on-path/scripts/install.ps1
@@ -1,4 +1,0 @@
-$path = [System.Environment]::GetEnvironmentVariable('PATH', 'User')
-$path = ($path.Split(';') | Where-Object { $_ -ne "$PSScriptRoot\bin-win" }) -join ';'
-$path = $path, "$PSScriptRoot\bin-win" -join ';'
-[Environment]::SetEnvironmentVariable('PATH', $path, 'User')
